HORNETQ-1573 Some Rest tests cannot build with JDK 1.6
The RestDeserializationTest was introduced in commit
b72408fb0e19db227a2bae6fe590f9c7466a2d65
It depends on Jetty which doesn't compile with
JDK 1.6 (class format is newer).

HornetQ should be able to compile and run with JDK 1.6.

The solution is to move the test into a separate module
and activated with proper JDK versions.